+++ torchdr/spectral_embedding/__init__.py
@@ -1,0 +1,6 @@
+from .pca import PCA
+from .kernel_pca import KernelPCA
+from .incremental_pca import IncrementalPCA
+from .phate import PHATE
+
+__all__ = ["PCA", "KernelPCA", "IncrementalPCA", "PHATE"]
