+++ torchdr/spectral_embedding/phate.py
