--- torchdr/tests/test_phate.py
@@ -1,42 +1,0 @@
-from torchdr.tests.utils import toy_dataset
-from torchdr.diffusion_embedding.phate import PHATE
-from torchdr.affinity.knn_normalized import NegPotentialAffinity
-from torchdr.utils import pykeops
-import torch
-import pytest
-
-DEVICES = ["cpu"]
-if torch.cuda.is_available():
-    DEVICES.append("cuda")
-USE_KEOPS = [False]
-if pykeops:
-    USE_KEOPS.append(True)
-
-
-@pytest.mark.parametrize("device", DEVICES)
-@pytest.mark.parametrize("backend", ["keops", "faiss", None] if pykeops else ["faiss", None])
-def test_potential_dist(device, backend):
-    data, _ = toy_dataset()
-    data = torch.tensor(data, dtype=torch.float32)
-    data.requires_grad = True
-    neg_affinity = NegPotentialAffinity(backend=backend, device=device)(data)
-    assert neg_affinity.shape == (data.shape[0], data.shape[0])
-    assert neg_affinity.min() < 0
-    neg_affinity.sum().backward()
-    assert data.grad is not None
-
-
-@pytest.mark.parametrize("device", DEVICES)
-@pytest.mark.parametrize("backend", ["keops", "faiss", None] if pykeops else ["faiss", None])
-def test_phate(device, backend):
-    torch.autograd.set_detect_anomaly(True)
-    data, _ = toy_dataset()
-    data = torch.tensor(data, dtype=torch.float32)
-    data = data.to(device)
-    phate = PHATE(backend=backend, device=device)
-    embedding = phate.fit_transform(data)
-    assert embedding.shape == (data.shape[0], 2)
-
-
-if __name__ == "__main__":
-    pytest.main([__file__])
