+++ torchdr/utils/__init__.py
@@ -10,7 +10,6 @@ from .geometry import (
     LIST_METRICS_KEOPS,
     LIST_METRICS_FAISS,
     pairwise_distances,
-    symmetric_pairwise_distances,
     symmetric_pairwise_distances_indices,
 )
 from .keops import LazyTensor, LazyTensorType, is_lazy_tensor, pykeops
@@ -89,7 +88,6 @@ __all__ = [
     "l2_loss",
     "check_NaNs",
     "pairwise_distances",
-    "symmetric_pairwise_distances",
     "symmetric_pairwise_distances_indices",
     "LIST_METRICS_KEOPS",
     "LIST_METRICS_FAISS",
